Fix selectors (classes) of notification badges elements
Franz was unable to show notification badges when there where unread
messages. It happened because the CSS classes used to lookup for
elements of chats where different from current values used by
chat.google.com.

This commit fixes the classes used as selectors to find chats in the
list with the _unread_ badge. It also extract the selectors as
variables, so it is easier to understand and maintain the code.