Fix selection of element for 'muted' notification
While setting up unread messages badge, the webview also checks if a
given chat is also 'muted'. If it is, the counter skips it.To check if
a conversation is muted, it gets the root element of the chat list item,
and looks up for an element with the muted class.

Inspecting the DOM, the current root element for a listed conversation
(contact or room) is a `span`. However, the code was using
`node.closest('content[role=listitem]')` to find out the parent element.
That always returns `undefined`. Consequently, the chained call to
`.querySelector(mutedSelector)` raises an Exception. Such exception does
not appear in the console, silenced some how, but the method exits right
there. Thus, the increment of notification counters (`directCount += 1`
and `indirectCount += '`) are not executed. Therefore, no notification
badge is set for the service.

This commit fixes the chat item (contact or group) parent element lookup
by fixing the element type `content` -> `span`. Hence, correctly sets
the notification badge.